Created protocol VideoAnalyzer to separate different types of analysis. Decoupled human results from MotionChart and used separate charts for each analysis result. Split main content view into several smaller views. Simplified video analysis error types. Added and restructured files as needed. Made optional values non-optional for various video attribute State vars. Various changes to improve performance, code structure and readability.

+++ MotionDetector/Video Analysis/MotionResult.swift
@@ -6,7 +6,7 @@ struct MotionResult: Identifiable {
     let time: Double
 
     /// The amount of difference between the frame at this time and the previous extracted frame.
-    let amount: Float
+    let amount: Double
 
     // MARK: - Identifiable conformance
